🚚 move: Header device 폴더 분리
